+++ CHANGELOG.md
@@ -14,6 +14,7 @@ All Notable changes to `bakame/tokei` will be documented in this file.
 - `IntervalSet::previous`
 - `IntervalSet::nearest`
 - `Interval::fromFormat`
+- `Interval::roundTo`
 - `Duration::fromDateInterval`
 - `Duration::fromFormat`
 - `Duration::format`